Added --run-command option, obsoletes --wrapper
Wrapper was an idea exclusively useful for running the binary via
i3-exec. While that is of utmost importance, the current --run-command
feature subsets wrapper and a *lot more*.

The current syntax is
```
	--run-command "/usr/bin/anybinary {}"
```

The --wrapper *could* be replaced with a similar structure
```
	--run-command "wrapper_binary '{}'"
```

I strongly advocate for `--run-command` and obsolete `--wrapper` as it is
way more awesome. The idea was directly inspired by rofi's -run-command