Move read() into constructor in Application
The constructor should be the thing that makes the class, that
initializes all variables. Application now uses exceptions to indicate
error while constructing.

When Application figures out that the desktop file is Hidden, is
NoDisplay or doesn't match OnlyShowIn and NotShowIn then there's no
reason to continue parsing it because it's useless for j4dd, so this
commit throws a disabled_error exception and destroys Application when
this happens.