ApplicationRunner: modify escape logic
This commit changes the escaping logic because the previous commit
changed how string values are expanded in Appplication.

The escaping is now done by the shell like Qt/KDE does it. The field
codes are now being escaped too to better comply with the standard.

This commit completely removes the temporary script creation for
Terminal=true programs. do_dmenu() now passes the program directly
to the terminal emulator.