Switch to unity builds
The Test.cc file was a more explicit way to do unity builds. By doing it
in CMake, tests can be built "normally" by changing a single line in
CMakeLists.txt. This could be also set up to make it user configurable
by passing arguments to the cmake executable, but that has been
overwritten in the CMakeLists.txt file.