Refactor AppManager + remove lookup functionality
AppManager can not handle lookup because the name list provided to and
from dmenu is formatted. The formatted name returned from dmenu could be
converted back to a "clean" name, but that is unreliable. There could be
more complex formatters implemented in the future. Application lookup
will now be responsibility of main.cc. AppManager now provides a name to
Application* mapping which makes this possible.