main: Fix bug in wait_on mode
poll() may return both POLLIN and POLLHUP at the same time under high
system load. Old implementation first checked for POLLHUP, whose handler
replaces fd with a new one. This breaks the POLLIN handler, so the old
implementation called `continue` to skip the POLLIN handler. Because the
POLLHUP handler completely replaces the fd, next iteration of poll()
can't register the POLLIN handler, so it will be ignored (!) if both
POLLIN and POLLHUP register at the same time.

Losing POLLIN events is not acceptable. This commit changes the order
of operation to first handle POLLIN and then POLLHUP.

This bug was likely introduced by
8ce1fbbe9e32212a66f0dcc620d71c6b246a45c3