Main: Allow explicitly overriding I3SOCK
Overriding i3 socket path was fully supported before (as documented in
the manpage). The mechanism relied on propagating I3SOCK environmental
variable to i3 executable which would then acknowledge it in
--get-socketpath. This can be troublesome when using other window
managers compatible with i3 IPC (like Sway). It introduces an arbitrary
requirement on i3 executable which is unnecessary because it's only used
for `i3 --get-socketpath`.

j4-dmenu-desktop now looks for I3SOCK and it handles it itself. When
$I3SOCK is set, i3 (nor sway) is executed whatsoever.

related to #166