Execute Exec manually instead of through shell
This is safer and it removes the middle man (/bin/sh). This change was
made because the latest specification is missing the [ character in the
reserved characters section of the Exec key specification. [ has special
meaning in dash, but the spec doesn't require it to be quoted. Passing
Exec to the shell is incorrect because of this (the shell approach could
be fixed, but this solution is better).