Ditch Bats and use pytest instead
I believe that this change will contribute to the usefulness and
maintainability of system tests.

The current implementation still has room for improvement. I will try
to refactor things when necessary.

This commit replaces Bats with pytest and nothing else. The test data
and other stuff still references bats. These references will be fixed
in other commits.

related to #164