Modify --desktop-file-quirks handling
I still don't find the handling of quirks "perfect", but users will
likely not have to configure it all that much. Everything is enabled
by default (for now) and there are only two modes to configure (for
now).

I have designed this system to resemble `mount -o` syntax. One
difference is that multiple --desktop-file-quirks specified in a single
j4dd invocation do not stack.